+++ src/Controller/Ffp3/OutputController.php
@@ -153,6 +153,11 @@ class OutputController
         return $this->handleToggle($request, $response);
     }
 
+    public function toggleOutputS3Test(Request $request, Response $response): Response
+    {
+        return $this->handleToggle($request, $response);
+    }
+
     private function handleToggle(Request $request, Response $response): Response
     {
         $params = RequestHelper::extractParams($request);